% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_zi_nb.R
\name{smle_zi_nb}
\alias{smle_zi_nb}
\title{Sieve maximum likelihood estimation for zero-inflated negative binomial regression problems with covariate measurement error}
\usage{
smle_zi_nb(
  analysis_formula,
  error_formula,
  data,
  no_se = TRUE,
  pert_scale = 1,
  tol = 1e-04,
  max_iter = 1000,
  output = "coeff"
)
}
\arguments{
\item{analysis_formula}{analysis model formula (or coercible to formula), a formula expression as for other regression models. The response should be the negative binomial model outcome, and the covariates for the zero inflation model should be included after a \code{|} as described in the documentation for \code{pscl::zeroinfl}.}

\item{error_formula}{formula, covariate error model formula (or coercible to formula), a formula expression as for other regression models. The response should be the error-free version of the error-prone of the covariate, and the covariate should be the names of the B-spline columns.}

\item{data}{dataset containing at least the variables included in \code{error_formula} and \code{analysis_formula}.}

\item{no_se}{Indicator for whether standard errors are desired. Defaults to \code{no_se = FALSE}.}

\item{pert_scale}{Size of the perturbation used in estimating the standard errors via profile likelihood. If none is supplied, default is \code{pert_scale = 1}.}

\item{tol}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{max_iter}{Maximum number of iterations allowed in the EM algorithm.}

\item{output}{character, level of fitted model output to be returned. Defaults to \code{output = "coeff"}, but \code{output = "all"} is also possible.}
}
\value{
\item{coefficients}{dataframe with final coefficient and standard error estimates (where applicable) for the analysis model.}
\item{vcov}{variance-covariance matrix for \code{coefficients} (where applicable).}
\item{converged}{indicator of EM algorithm convergence for parameter estimates.}
\item{se_converged}{indicator of standard error estimate convergence.}
\item{converged_msg}{(where applicable) description of non-convergence.}
}
\description{
Sieve maximum likelihood estimation for zero-inflated negative binomial regression problems with covariate measurement error
}
